+++ lib/core/entities/entities.dart
@@ -3,7 +3,7 @@ export 'custom_list.dart';
 export 'game_object.dart';
 export 'game_state.dart';
 export 'health_management.dart';
-export 'invader_data_model.dart';
+export 'matrix8x12.dart';
 export 'particle.dart';
 export 'score.dart';
 export 'setting.dart';
