+++ lib/feature/on_play/models/models.dart
@@ -1,5 +1,6 @@
 export 'bullets.dart';
 export 'health_management.dart';
+export 'invader_matrix.dart';
 export 'player_ship_body.dart';
 export 'player_shoot.dart';
 export 'score_manager.dart';
