+++ lib/feature/on_play/utils/paints/enemy_ships_paint/invader_paint_a.dart
@@ -4,7 +4,7 @@ import '../../../../../core/entities/entities.dart';
 
 //* based Invader ship is 12x8 bloc
 class InvaderPaintA extends CustomPainter {
-  final InvaderMatrix invaderMatrix;
+  final Matrix8x12 invaderMatrix;
 
   InvaderPaintA({
     required this.invaderMatrix,
@@ -15,11 +15,11 @@ class InvaderPaintA extends CustomPainter {
     // InvaderMatrix.printData(invaderMatrix);
 
     Size blocSize = Size(
-      size.width / InvaderMatrix.col,
-      size.height / InvaderMatrix.row,
+      size.width / Matrix8x12.col,
+      size.height / Matrix8x12.row,
     );
-    for (int i = 0; i < InvaderMatrix.row; i++) {
-      for (int j = 0; j < InvaderMatrix.col; j++) {
+    for (int i = 0; i < Matrix8x12.row; i++) {
+      for (int j = 0; j < Matrix8x12.col; j++) {
         canvas.drawRect(
           Rect.fromLTWH(
             j * blocSize.width,
@@ -28,7 +28,7 @@ class InvaderPaintA extends CustomPainter {
             blocSize.width - 2,
             blocSize.height - 2,
           ),
-          Paint()..color = InvaderMatrix.blocColor(invaderMatrix.data[i][j]),
+          Paint()..color = Matrix8x12.blocColor(invaderMatrix.data[i][j]),
         );
       }
     }
