+++ lib/main.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'app.dart';
 import 'core/entities/entities.dart';
+import 'feature/on_play/models/models.dart';
 import 'feature/on_play/utils/paints/enemy_ships_paint/enemy_ships_paint.dart';
 
 void main() {
@@ -42,6 +42,7 @@ class _testPaintState extends State<_testPaint>
   late Animation<double> animation =
       Tween<double>(begin: 0, end: 1).animate(controller);
 
+  bool isXState = true;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -49,6 +50,12 @@ class _testPaintState extends State<_testPaint>
       body: Center(
         child: Column(
           children: [
+            ElevatedButton(
+                onPressed: () {
+                  isXState = !isXState;
+                  setState(() {});
+                },
+                child: Text("Change State")),
             ElevatedButton(
                 onPressed: () {
                   controller.forward();
@@ -62,11 +69,14 @@ class _testPaintState extends State<_testPaint>
             Container(
               // color: Colors.cyanAccent.withOpacity(.3),
               decoration: BoxDecoration(border: Border.all(color: Colors.red)),
-              width: InvaderMatrix.col * size,
-              height: InvaderMatrix.row * size,
+              width: Matrix8x12.col * size,
+              height: Matrix8x12.row * size,
               child: CustomPaint(
                 size: Size(size, size),
-                painter: InvaderPaintA(invaderMatrix: InvaderMatrix.aY),
+                painter: InvaderPaintA(
+                  invaderMatrix:
+                      isXState ? InvaderMatrixA.x() : InvaderMatrixA.y(),
+                ),
               ),
             ),
           ],
